requirements.txt
fastapi
uvicorn[standard]
asyncpg
pandas
jinja2
python-multipart
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt
